# docs/dz8/swagger.yaml
openapi: 3.0.0
info:
  title: Микросервисная платформа электронной коммерции
  description: |
    API для микросервисной платформы электронной коммерции, включающей сервисы заказов, 
    оплаты, доставки, склада, биллинга и уведомлений
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Сервис заказов
  - url: http://localhost:8081
    description: Сервис биллинга
  - url: http://localhost:8082
    description: Сервис уведомлений
  - url: http://localhost:8083
    description: Сервис платежей
  - url: http://localhost:8084
    description: Сервис склада
  - url: http://localhost:8085
    description: Сервис доставки
...
paths:
...
  /api/v1/auth/register:
    post:
      tags:
        - Auth
      summary: Регистрация нового пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
                first_name:
                  type: string
                last_name:
                  type: string
              required:
                - username
                - email
                - password
      responses:
        '201':
          description: Пользователь успешно зарегистрирован
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  username:
                    type: string
                  email:
                    type: string
        '400':
          description: Ошибка в запросе
  /api/v1/auth/login:
    post:
      tags:
        - Auth
      summary: Аутентификация пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                  format: password
              required:
                - username
                - password
      responses:
        '200':
          description: Успешная аутентификация
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    type: object
                    properties:
                      id:
                        type: integer
                      username:
                        type: string
        '401':
          description: Неверные учетные данные

  # Сервис заказов - Заказы
...
  /api/v1/accounts:
    post:
      tags:
        - Billing
      summary: Создание аккаунта для биллинга
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: integer
                email:
                  type: string
                  format: email
      responses:
        '201':
          description: Аккаунт создан
        '400':
          description: Ошибка в запросе
  /api/v1/accounts/{user_id}:
    get:
      tags:
        - Billing
      summary: Получение информации об аккаунте пользователя
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Информация об аккаунте
        '404':
          description: Аккаунт не найден
  /api/v1/billing/account:
    get:
      tags:
        - Billing
      summary: Получение информации о своем аккаунте
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Информация об аккаунте
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  user_id:
                    type: integer
                  balance:
                    type: number
                    format: float
        '401':
          description: Пользователь не авторизован
  /api/v1/billing/deposit:
    post:
      tags:
        - Billing
      summary: Пополнение баланса
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: number
                  format: float
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Баланс пополнен
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  transaction:
                    type: object
        '400':
          description: Ошибка в запросе
  /api/v1/billing/withdraw:
    post:
      tags:
        - Billing
      summary: Снятие средств с баланса
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: number
                  format: float
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Средства сняты
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  transaction:
                    type: object
        '400':
          description: Ошибка в запросе или недостаточно средств

  # Сервис платежей
...
  /api/v1/notifications:
    post:
      tags:
        - Notifications
      summary: Отправка уведомления
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: integer
                type:
                  type: string
                  enum: [email, sms, push]
                title:
                  type: string
                message:
                  type: string
                email:
                  type: string
                  format: email
      responses:
        '201':
          description: Уведомление отправлено
        '400':
          description: Ошибка в запросе
    get:
      tags:
        - Notifications
      summary: Получение списка всех уведомлений
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Список уведомлений
          content:
            application/json:
              schema:
                type: object
                properties:
                  notifications:
                    type: array
                    items:
                      type: object
                  total:
                    type: integer
  /api/v1/notifications/{id}:
    get:
      tags:
        - Notifications
      summary: Получение уведомления по ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Информация об уведомлении
        '404':
          description: Уведомление не найдено
  /api/v1/users/{id}/notifications:
    get:
      tags:
        - Notifications
      summary: Получение списка уведомлений пользователя
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Список уведомлений
          content:
            application/json:
              schema:
                type: object
                properties:
                  notifications:
                    type: array
                    items:
                      type: object
                  total:
                    type: integer

  # Health checks
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...